Migrating to typescript and using new API
